separate out proof request in a separate file
